don't pin coverage to < 4.0.0, stop testing py3.2 (#95)
So the coverage dep was pinned to < 4.0.0 because of an issue
between coverage and Python 3.2 (py3.2 doesn't accept the
u'' syntax). This is fairly well-known, it seems, and the
standard response is simply to drop Python 3.2 testing:

https://github.com/geopython/OWSLib/issues/291
https://github.com/menegazzo/travispy/issues/20

ipython dropped Python 3.2 support for the same reason:

https://stackoverflow.com/questions/28724785

so I think this is a sensible choice here too. It's not
workable to pin coverage to an old and unsupported release
indefinitely, and causes problems for downstreams which
include coverage 4.x.